Reimplementing sync_publish
Cannot use the normal subscribe method, because after the response is
received the loop is not exited, and subscribe will continue listening
for new messages forever.

Also, in case the response is never received, the subscribe will
continue waiting and the code handling the timeout is never reached.

Provide a different sync_subscribe method that will poll for a single
response instead of looping and receiving constantly. Polling will also
handle the timeout.