Run subscription in its own thread
No more need for a user to start a separate thread to run async subscribe.

Introducing SubscriptionHandler class. This class will be returned for
every async subscription that will be used later to un-subscribe a
topic. Note that this will take care of stopping subscription internal
thread and perform zmq un-subscribe.