Fix sync_publish losing responses
Some responses can be very fast and sent back before the orchestrator
subscribes to it. Thus, a response is never received and the
sync_subscribe method throws a timeout exception.

Fix this by subscribing before publishing the request.

Note that the subscription has to be to the 'returnAddress', not to the
'topic'.