xMsgNode: hack shutdown hook to avoid ZMQ exceptions
Thread.stop() is not the recommended way to stop a thread, but it is
good enough to avoid the exception that would be thrown if the thread is
interrupted.

TODO: the exception is caused by the thread being interrupted,
so every call to a blocking ZMQ method should be redone and fixed on 'master',
to use the proper Thread.interrupt() call without problems.

zmq.ZError$IOException: java.nio.channels.ClosedByInterruptException
        at zmq.Signaler.recv(Signaler.java:163)
        at zmq.Mailbox.recv(Mailbox.java:99)
        at zmq.SocketBase.process_commands(SocketBase.java:830)
        at zmq.SocketBase.recv(SocketBase.java:714)
        at org.zeromq.ZMQ$Socket.recv(ZMQ.java:1257)
        at org.zeromq.ZFrame.recv(ZFrame.java:326)
        at org.zeromq.ZFrame.recvFrame(ZFrame.java:359)
        at org.zeromq.ZMsg.recvMsg(ZMsg.java:223)
        at org.zeromq.ZMsg.recvMsg(ZMsg.java:205)
        at org.jlab.coda.xmsg.xsys.regdis.xMsgRegistrar.run(xMsgRegistrar.java:125)