xMsg: update registration API
Define the following methods for registration of xMsg pub/sub actors:

- To send the request to the global front-end registrar:

    registerPublisher
    registerSubscriber
    removePublisher
    removeSubscriber

- To send the request to the local registrar:

    registerLocalPublisher
    registerLocalSubscriber
    removeLocalPublisher
    removeLocalSubscriber

Any request will register or remove a *local* actor (i.e. running in the
current node), with the specified global or local registrar.

Temporarily remove the host and port parameters from the requests.
Fill the registration with the default values. Handle a different proxy
configuration on next release.