xMsg: give a name to every actor
The semantic of the core xMsg class is to represent an xMsg actor, who
can register as a publisher/subscriber, discover other actors and
publish/subscribe messages.

By receiving a name when the actor is constructed, that name can be used
for registration purposes and as the sender of requests and messages.