xMsg: update registration/discovery API
Follow-up to commits 97b5caecbe and f6925ce6a4, that set the names of
the methods, and to db35c920f0 and b8ffafe998, that changed the
signatures to use xMsgTopic.

This commit removes the "name" or "sender" parameter from the
signatures.

xMsg has a "name" field now to identify the actor, by 71b17a1235.
It is the xMsg actor who is registering as a publisher or subscriber,
so its name should be used when sending the registration.