Refactor subscription handler
Rename SubscriptionHandler to xMsgSubscription (not just an OCD change,
but to be consistent with all classes starting with xMsg).

Call Socket.subscribe in the handler (to be consistent with the call to
Socket.unsubscribe also in the handler). Also, get the message and pass
it to the handle method (keep the usage of the connection in the
handler).

Update permissions of the class methods. Give them package access only.
This means that only the xMsg class can start and stop a subscription.

The subscription thread will have a custom name, for identification
purposes.