Fix unsubscription
The subscription runs in a background thread, and this thread may never
finish by calling unsubscribe, because the thread is always waiting for
a message to arrive before returning to the loop.

Replace the blocking recv method will poll and a timeout. Poll will just
return if no message arrived. This allows to periodically check the
control variable to exit the receiving loop and stop the subscription.