connect: give time to stablish connection
Add an overload to wait after a new a connection is created, to have
time to establish the connections and allow the PUB socket to receive
the list of subscriptions from the XSUB socket in the proxy.

If there is no delay between the connect call and the publish call, the
PUB socket will drop messages if the expected subscription is not
received yet.