Design update: xMsgMessage contains only raw data
Change the design of the message class to only contain raw binary data
instead of any object.

xMsg will not handle serialization/deserialization anymore, but send
whatever bytes are given, without caring about what is in the data. It
is now responsibility of the user to do the serialization and set the
proper mime-type of the data. xMsgData is still used as a helper class
to provide serialization for primitive data.

With this update, an xMsg data message will be composed of three frames:
topic, metadadata, and data. There is no frame with information about if
the data is serialized or not, because it is expected to be raw bytes.
The mime-type on the metadata should be used to identify what's in the
data.