xMsgRegDriver: revert removeAll to previous spec
The original spec is to remove from a global registrar service all
actors running on a given host. It is mainly used to remove the
registration of a node being shutdown.

Current spec removes all actors running in the same host as the
registrar service, which is not very useful.

A request to totally clean all registration for a given registrar is
still missing from both driver and database.

This change also fixes the Registrar integration test.