xMsg: fix generated replyTo topic on syncPublish
A random number between 0 and 100 is creating many conflicts when a
unique replyTo topic must be generated, and those response messages are
being lost.

Replace the random number generation with an identifier plus a unique
sequence of 6 digits:

    ret:<local_address>:<prefix_digit><6_digits>

The identifier is the local proxy address used by the xMsg agent.
Thus, it will be the same for every request sent by the agent.
This ensure actors in different nodes will have a different identifier.

The sequence of digits is global in the JVM. To ensure every single
requests uses a different topic, every time a replyTo is required, the
next number in the sequence will be used. The sequence goes from 1000000
to 1999999 and then repeats. 6 digits seems good enough, because 1
million of concurrent syncPublish messages is not even considered.