Receive sync publish response on same thread
Creating a thread is expensive and requires synchronization to check if
a response was received. And there is actually no need to start a
background thread to wait for the response.

Use a subscription object that does not run a thread, and
replace the thread with a simple while loop that polls the SUB socket
for a response.

This also fixes a bug with the temporary subscription not being
unsubscribed.