Receive sync publish response on same thread
Creating a thread is expensive and requires synchronization to check if
a response was received. And there is actually no need to start a
background thread to wait for the response.

Use a subscription object that does not run a thread, and
replace the thread with a simple while loop that polls the SUB socket
for a response.

This also fixes a bug when an actor is running many sync publication
requests on different threads. As the response subscriptions are running
on background threads, the responses will be processed by the actor's
threadpool, which will be at full capacity, putting responses on queue,
and thus resulting in timeout of some of the requests.