perf: workaround for linger regression
96c29cc4f310250faa0e5311fb1af51e720e456b introduced a connection factory
which is not calling context.destroy() anymore. That prevented the
ZMQ context of RemoteThroughput to send all the messages on queue before
ending, and therefore LocalThroughput never finishes.

As a workaround, ensure calling context.destroy(), which does use the
linger option to wait until all messages are sent before destroying the
context.

A better solution is needed, though. Destroying an actor requires
careful thinking to ensure only sockets related to the actor are
destroyed and no message is lost.