proto: remove default value of replyTo metadata
Having "undefined" as default value just complicates the clients using
the API, because then it is required to check twice for a replyTo
address: check if the field is present and check if the field is not
"undefined".

Just leave it empty when is not defined. It is easier to check.