Connection: filter control messages on subscription handler
If an actor A starts a new subscription but reusing a topic of an already
running subscription of another actor B, the control message will be
published to both actors. Then actor B will fail to parse the message
because it is not a data message (and it only has two frames).

Preven this by filtering control messages before parsing the raw message.

Note that it is unlikely to receive control messages, though. Branch
prediction should be smart enough to preload the normal data message
path.