subscription: safe publish sub object to thread
Make sure it is safe to use the SUB socket in the background thread,
and it is only used by the thread (thus, unsubscription is done at the
end of the run method).

Subscription is done before creating the background thread, to notify
of any error to the current thread (i.e. the thread should not be
created if the subscription could not start).