connection: use a socket factory
Put all code related to creating and connecting sockets into a single
class, instead of being duplicated all over the place.

Note that there is a single place to create a socket now, and by default
the sockets will have a HWM of 0, so there is no need to set the HWM as
the default action of the connection setup interface.