subscription: simplify publication of socket
No need to pass the socket to the constructor of the runnable object and
store it in a final field to ensure safe publication.

The Thread.start() call happens-before any action in the started thread,
so any modification to the socket state done before starting the
background polling thread will be visible inside the thread.

Thus, it is safe to subscribe the socket in the parent thread and then
start receiving messages in the background thread.

We want to subscribe in the parent thread to catch and report any
subscription problem. The background thread will not be started if
subscription failed.