subscription: subscribe when calling start()
Do not subscribe on the constructor of the subscription handler.
It makes complicated to have a concurrent map of subscriptions, where
two threads may try to start the same subscription (it is an error to do
so, but still a possibility), and this may create two objects trying to
be saved into the map.

By subscribing on the start() method, we can have the handlers created
(and discarded) without subscribing the socket and blocking to check
that the subscription is receiving messages.

The subscription can then start only if the handler was successfully put
into the map.