connection: destroy all connections individually
Do not close the connections through the context and the factory,
but close each connection one by one.

Depending on the context to destroy sockets is not a good idea
because xMsg connections are distributed through many threads,
so they must be closed by the thread that is using them,
or obtained from the thread-safe connections pool before closing them.
Otherwise the context will certainly hang when it tries to close the
sockets that it still manages.