connection: add helper to close sockets after errors
When creating many sockets, if one of them fails the ones that got
created must be destroyed.

The common pattern is to have a null variable to which the new socket
will be assigned, so to know if the socket can be closed there must be a
null check.

Add a helper method that encapsulates this check and simplifies the
code.