Reimplement sync-publication
Use a single poller thread to wait for response messages instead of
creating a new subscription for every request.

Keep a single connection per proxy. Thus, two concurrent sync
publication requests to the same proxy will receive the response through
the same background connection. The connection will be created and
subscribed the first time a request is sent to the given proxy.

The poller thread will loop through all connections to check for new
messages. All received messages will be stored in a map, with the topic
as the key.

The topic is still unique for every sync-publish request. Thus, waiting
for a response is now implemented as periodically checking the map for a
message with same topic of the request.