Revert "Destroy context on exit"
This reverts the global shutdown hook added by commit
02f9ef4813deac643ec6af4f1a6d67a9628967a5.

This creates problems for applications that were closed just fine before
and now hang because the sockets may still be open (the context doesn't
manage the sockets anymore, and the xMsg actor must be destroyed to
ensure all sockets are closed).

For example: having an actor publishing messages in an infinite loop,
without a shutdown hook to close the actor, will hang when CTRL-C is
received because the context will be destroyed (it is registered as a
shutdown hook) but all the sockets are still open.

It is preferable to have an application that closes without hanging, and
that doesn't require extra handling from the client, like needing a
shutdown hook for the basic case of interrupting the application. Thus,
forcing a shutdown hook to always destroy the context is not a good
idea.