connection: wrap proxy connection
Keep the real connection to a proxy as internal implementation, and
create a wrapper class to hide the sockets and low level access from
xMsg users.

Now there is symmetry between proxy and registrar connections. Both
classes are called "Driver", are managed by the connection pool and are
hidden from the user.

The wrapper class will be also useful to automate returning the proxy
connection to the pool.