connection: make wrapper closeable
Simplify usage and correctness of connections returned by the connection
pool.

All connections must be returned to the pool, to be closed when the
actor is destroyed. Forcing the user to manually return the connection
on a finally block is error prone, and too verbose.

With the connection wrapper implementing Closeable, the new
try-with-resources block can be used, the close() method will return the
internal proxy connection to the pool, and IDEs can show a warning if
the wrapper is not closed.