xMsg: subscribe to address, not to connection
When the subscription is not to the default proxy, a connection must be
given. The subscription will take control of the connection in a
background thread. But the connection class is closeable now (to ensure
publication returns the connection to the thread pool) and most IDEs
will complain than the connection is not closed after being passed to
the subscription.

Suppressing the warning or closing the connection (that should be empty
because the real proxy connection is used by the subscription thread)
is just adding complications to the clients. Prefer to pass an address
to avoid confusing clients, and handle the connection internally.