xMsg: add publish to address
Using a connection object requires getting a connection every time, and
ensuring it is closed (returned to the pool). Since the logic is always
the same, add an option to use just an address, to simplify the code of
the clients. Obtaining and releasing the connection can be done
internally. It also keeps symmetry with the subscribe method, that uses
an address now.

The publish overload receiving a connection should be used for
long-running publications to the same proxy, where there is no need to
return the connection to the pool every time.