xMsg: remove getConnection overload
Remove the getConnection overload that receive the host address as an
string. This is just to keep the same structure for all methods:

- no address parameter (default address)
- specific address object

The pub/sub methods and the reg/disc methods do not accept an string as
the address.

The choice is: all methods provide an string overload, or none.
Prefer the latter, for consistency.