scripts: cache classpath instead of using Gradle
Commit bf926eed8336 ("scripts: use classpath instead of Gradle
JavaExec") started using Java to run development scripts instead of the
Gradle execution task, because the startup was too slow and the Gradle
daemon was killed when using Ctrl+C.

But Gradle knows about the dependencies so a task was used to obtain the
proper classpath. The startup time was better, but still slow compared
to running just Java.

To avoid calling the still not-fast Gradle task every time to generate
the development classpath, use a custom task to cache the classpath into
a file in the build directory, that runs when the project is built.
If dependencies change, the file will be updated accordingly.

Now the development scripts can just read the classpath from the file
instead of using a Gradle task, and they run as fast as the installed
version of the scripts.

Hooray.