Update Gradle wrapper version to 2.14.1
Since CUE systems cannot download the Gradle wrapper due to SSL
restrictions, bump the Gradle wrapper version to the same version used
by the other projects, so a proper binary distribution of Gradle can be
downloaded and used with all current projects.

Normally a newer Gradle version would work just fine with scripts
created for older versions, but the current protobuf plugin used by xMsg
only works with Gradle 2.11 or lower.

So, in order to use the same newer Gradle version for all projects, the
wrapper and the protobuf plugin need to be updated.