registration/database: add a new indirection level
Some applications, like CLARA, contain a high number of actors, all
with unique topics. When this happens, querying the database requires to
check all the topics for just a few matches.

To speed up finding actors, add a new top-level to group them, based in
the domain (which is probably the one value that several actors will
share within their unique topics).

Now the in-memory database looks like:

    Map { domain :
          Map { Topic :
                Set [ Registration ]
              }
        }

This allows searching only a subset of actors for matching topics or
query terms, while ignoring all the others registered actors.