subscription: don't store the handler in case of errors
If the subscription failed to start, then the handler should not be
stored in the subscriptions map, and the connection should be closed.

One option is to start the subscription after the handler was
successfully stored in the map, catch any errors and remove the handler
from the map in that case, and rethrow.

But that alternative adds too much lines of code. The simpler one is to
always start the subscription. The probability of two subscriptions
racing each other to be registered in the map is low. If that happens,
then the subscription that lost the race must be stopped.

The existing catch block ensures the connection is closed in case of any
error or duplication.