subscription: support multiple topics
0MQ allows subscription to multiple topics. xMsg should provide the
same, to have more flexible applications.

Add the necessary overloads to the API to accept a set of topics to be
subscribed, along with the single topic API.

The subscription will subscribe the socket with all the topics, but to
check if the connection is up only one of the topics is used for the
control messages from the proxy.