subscription: support single-thread subscriptions
Add an option to change the subscriptions to run the callbacks in
single-thread mode, where all messages are processed sequentially by the
receiving thread itself. This is, the subscription thread will be used
to process the callback instead of delegating the task to the thread
pool, and it will not receive another message until the callback for the
current message is finished.

Keep the default mode to multi-thread (i.e. using the thread pool), but
give the option to clients that want a bit more control (for example,
sharing an external thread-pool between actors).