address: fix fallback to loopback address
Redo 3e8cc6b11e88 ("address: use loopback if no interfaces were found")
and do not call InetAddres.getLocalHost() when there was no non-loopback
address in all the network interfaces. It was expected that the method
would return the IPv4 loopback since no other interface was available,
but it could also return the IPv6 loopback (or the link-local
equivalent, which fails the isLoopbackAddress() method).

Instead, cache the IPv4 loopback address when iterating through the
interfaces. Throw only if there was no IPv4 loopback in all the
interfaces.