build.gradle: fix protoc version on older systems
The protoc artifact downloaded by the protobuf plugin may fail on older
Linux systems with /lib64/libc.so.6: version `GLIBC_2.14' not found.

Add an option to ignore the downloaded compiler and use a local protoc
binary with a Gradle property. (Multiple assignments are allowed in the
protoc block. The last one wins.)