registrar: fix filter by host
When a filter to return all the actors of a given host is received, the
port in the protobuf query was hardcoded to the default port, thus only
some of the actors were returned, and those with different ports were
ignored.

Fix by setting the default port value to 0 and filtering the port only
if it set to a different value.