context: add factory method to create new contexts
It may be necessary to create new context wrappers to use instead of the
global singleton. This contexts should be used when the 0MQ context has
to be destroyed before exiting the process (the global context must be
destroyed at the very end, after all actors and connections have been
destroyed).