message: parse data using the mime-type
When returning an object without specifying the data type class, do not
use the Object.class parameter to deserialize the data using Java
serialization, because it would not work for primitive-types stored as
protobuf data.

For example, if msg contains an integer, parseData(msg) should return an
Object that can be cast to Integer, instead of failing.

So, use the mime-type from the metadata to parse the value with the
proper Java type. Only if the mime-type is not one of the default xMsg
types, use Java serialization to obtain an Object.