Add settings.gradle file
It is recommended by the official documentation that every project
should contain a settings.gradle file [0]. Use this file to define the
project name as "xmsg", which otherwise would be called as the parent
directory, most probably "xmsg-java". With the project name set, there
is no need to set the archivesBaseName property.

Having the project name set also fixes using a composite build [1] in
Clara with xMsg as included build, for example:

    $ ./gradlew --include-build ../xmsg-java build

where the substitution requires the project name to be "xmsg", or it
will not work.

[0]: https://docs.gradle.org/current/userguide/organizing_gradle_projects.html#always_define_a_settings_file
[1]: https://docs.gradle.org/current/userguide/composite_builds.html